refactor: extract walker specific logic to `walker.c`
