+++ src/ecs/entities.h
@@ -4,5 +4,5 @@
 
 EntityBuilder PlayerCreate(f32 x, f32 y);
 EntityBuilder ECreateBlock(f32 x, f32 y, f32 width, f32 height);
-EntityBuilder ECreateWalker(f32 x, f32 y);
+EntityBuilder WalkerCreate(f32 x, f32 y);
 EntityBuilder ECreateCloudParticle(f32 centerX, f32 centerY, Vector2 direction);
